requirement.txt
streamlit
pdfplumber
sentence-transformers
scikit-learn
nltk
beautifulsoup4
requests
pandas
sounddevice
scipy
torch
transformers
groq
googlesearch-python
